fp fix

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,10 +11,11 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	size_t i;
+
 	if (array != NULL && size > 0 && action != NULL)
 	{
-		for ( i = 0; i < size; i++)
+		for (i = 0; i < size; i++)
 		{
 			action(array[i]);
 		}
